Max pomodoro

+++ smartPromise.sol
@@ -3,13 +3,12 @@
 pragma solidity ^0.8.3;
 
 contract smartPromiseContract {
-    
     struct promiseData {
         address initialDepositor;
         uint256 promiseCollateral;
         string promiseTitle;
-        uint promiseIdentifier;
-        uint promiseAcceptDeadline;
+        uint256 promiseIdentifier;
+        uint256 promiseAcceptDeadline;
         address[] promiseParticipators;
     }
 
@@ -20,33 +19,47 @@ contract smartPromiseContract {
         newPromise.initialDepositor = msg.sender;
         newPromise.promiseCollateral = msg.value;
         newPromise.promiseTitle = _promiseTitle;
-        newPromise.promiseIdentifier = uint(keccak256(abi.encodePacked(
-            block.difficulty, block.timestamp, block.coinbase))) % 2**160; 
-            //generates a random number to use as a identifier
+        newPromise.promiseIdentifier =
+            uint256(
+                keccak256(
+                    abi.encodePacked(
+                        block.difficulty,
+                        block.timestamp,
+                        block.coinbase
+                    )
+                )
+            ) %
+            2**160;
+        //generates a random number to use as a identifier
         newPromise.promiseAcceptDeadline = block.timestamp + 10 minutes;
         newPromise.promiseParticipators = new address[](1);
         newPromise.promiseParticipators[0] = msg.sender; // want to add addresses to array promiseParticipator
         smartPromises.push(newPromise);
     }
 
-
-    function joinPromise(uint _promiseUID) public payable {
-        for (uint i = 0; i < smartPromises.length; i++) {
-            if (smartPromises[i].promiseIdentifier == _promiseUID &&
-            smartPromises[i].promiseAcceptDeadline > block.timestamp &&
-            smartPromises[i].promiseCollateral == msg.value) {
-                smartPromises[i].promiseParticipators.push(msg.sender); 
+    function joinPromise(uint256 _promiseUID) public payable {
+        for (uint256 i = 0; i < smartPromises.length; i++) {
+            if (
+                smartPromises[i].promiseIdentifier == _promiseUID &&
+                smartPromises[i].promiseAcceptDeadline > block.timestamp &&
+                smartPromises[i].promiseCollateral == msg.value
+            ) {
+                smartPromises[i].promiseParticipators.push(msg.sender);
             } else {
-                revert("Error: Invalid promise UID or deadline has passed or collateral does not match");
+                revert(
+                    "Error: Invalid promise UID or deadline has passed or collateral does not match"
+                );
             }
-        } 
+        }
     }
 
-    function endSmartPromise(uint _promiseUID) public payable {
+    function endSmartPromise(uint256 _promiseUID) public payable {
         require(address(this).balance > 0, "contract is empty");
-        for (uint i = 0; i < smartPromises.length; i++) {
+        for (uint256 i = 0; i < smartPromises.length; i++) {
             if (smartPromises[i].promiseIdentifier == _promiseUID) {
-                payable(msg.sender).transfer(smartPromises[i].promiseCollateral);
+                payable(msg.sender).transfer(
+                    smartPromises[i].promiseCollateral
+                );
             } else {
                 revert("Invalid promise identifier");
             }
